requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.3            # via django
click==7.0                # via pip-tools
django==3.0.2
djangorestframework==3.11.0
mysqlclient==1.4.6
pip-tools==4.4.0
pytz==2019.3              # via django
six==1.14.0               # via pip-tools
sqlparse==0.3.0           # via django
